Small API to force adapter change.
Using 'require' to change adapters works well, until you have to switch from
adapter 'a', 'b', and then 'a' again. Because require doesn't bother with
files it has already seen, the second call to require 'a' never runs.

This change allows the adapters to be changed using require, but it also adds
a small API for forcing changes. The benchmark task needed this, since it
has to rehearse each adapter, and then benchmark them.